Added a bunch of things
A working course info editor using Markdown to parse the text (heck
yeah). It stores previous versions of it to revert back to, thanks
github for the inspiration.

Added anthony's css to the login page which looks wicked, so thats nice

And just general steps forward in the way of a CMS